requirements/base.txt
argon2-cffi==23.1.0
celery==5.4.0
django==4.2.15
django-auth-ldap==4.8.0
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-cryptography==1.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django_filter==24.3
django-mass-edit==3.5.0
django-model-utils==5.0.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
drf-spectacular==0.27.2
python-ldap==3.4.4
flower==2.0.1
hiredis==3.0.0
Pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-slugify==8.0.4
pytz==2024.2
rcssmin==1.1.2
redis==5.0.8
requests==2.32.3
uvicorn[standard]==0.30.6
whitenoise==6.7.0
django-celery-results==2.5.1
matplotlib==3.9.2
numpy==2.1.1
